+++ libft/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: ylee <ylee@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/06 11:54:29 by ylee              #+#    #+#             */
-/*   Updated: 2020/10/12 14:40:56 by ylee             ###   ########.fr       */
+/*   Updated: 2020/10/13 13:03:12 by ylee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int	ft_strlcpy(char *dst, const char *src, size_t size)
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t	i;
